backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
sentence-transformers==2.2.2
google-generativeai==0.3.2
numpy==1.26.0
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.0
chromadb==0.4.18